+++ ContentPatcher/Framework/PatchManager.cs
@@ -149,12 +149,13 @@ namespace ContentPatcher.Framework
         /// <param name="weather">The current in-game weather (if applicable).</param>
         /// <param name="spouse">The current player's internal spouse name (if applicable).</param>
         /// <param name="dayEvent">The day event (e.g. wedding or festival) occurring today (if applicable).</param>
-        public void UpdateContext(IContentHelper contentHelper, LocalizedContentManager.LanguageCode language, SDate date, Weather? weather, string dayEvent, string spouse)
+        /// <param name="seenEvents">The event IDs which the player has seen.</param>
+        public void UpdateContext(IContentHelper contentHelper, LocalizedContentManager.LanguageCode language, SDate date, Weather? weather, string dayEvent, string spouse, int[] seenEvents)
         {
             this.VerboseLog("Propagating context...");
 
             // update context
-            this.ConditionContext.Set(language: language, date: date, weather: weather, dayEvent: dayEvent, spouse: spouse);
+            this.ConditionContext.Set(language: language, date: date, weather: weather, dayEvent: dayEvent, spouse: spouse, seenEvents: seenEvents);
             IDictionary<ConditionKey, string> tokenisableConditions = this.ConditionContext.GetSingleValueConditions();
 
             // update patches